requirements.txt
pyserial>=3.5
aiohttp>=3.8.0